fix circiular require bug
